requirements.txt
PyQt5==5.15.9
python-docx==0.8.11
PyPDF2==3.0.1
requests==2.31.0
beautifulsoup4==4.12.2
lxml==4.9.3 